Crate fifteen_min

Modules

amenities_details ğŸ”’

bus ğŸ”’

common ğŸ”’

from_amenity ğŸ”’

isochrone ğŸ”’

render ğŸ”’

score_homes ğŸ”’

single_start ğŸ”’

This is a tool to experiment with the concept of 15-minute neighborhoods. Can you access your daily needs (like groceries, a cafe, a library) within a 15-minute walk, bike ride, or public transit ride of your home?

Functions

main

run ğŸ”’

Type Aliases

App ğŸ”’