Module fifteen_min::find_amenities

Structs

FindAmenity
Calculate isochrones around each amenity on a map and merge them together using the min value

Results

Functions

create_multi_isochrone
For every one of the requested amenity on the map, draw an isochrone