Function fifteen_min::find_home::score_houses

fn score_houses(
    app: &SimpleApp<()>, 
    amenities: Vec<AmenityType>, 
    options: Options, 
    timer: &mut Timer<'_>
) -> HashMap<BuildingID, Percent>

For every house in the map, return the percent of amenities located within a 15min walkshed. A single matching business per category is enough to count as satisfied.