fifteen_min::score_homes
Function score_houses_by_one_match

fn score_houses_by_one_match(
    app: &SimpleApp<Options>,
    amenities: Vec<AmenityType>,
    timer: &mut Timer<'_>,
) -> (Counter<BuildingID>, MultiMap<AmenityType, BuildingID>)

For every house in the map, return the number of amenity types located within a 15min walkshed. A single matching business per category is enough to count as satisfied.