fifteen_min
Module single_start

This is a tool to experiment with the concept of 15-minute neighborhoods. Can you access your daily needs (like groceries, a cafe, a library) within a 15-minute walk, bike ride, or public transit ride of your home?

See https://github.com/a-b-street/abstreet/issues/393 for more context.

Structs

SingleStart

This is the UI state for exploring the isochrone/walkshed from a single building.

Functions

panel_contents 🔒