Module fifteen_min::viewer

This is a tool to experiment with the concept of 15-minute neighborhoods. Can you access your daily needs (like groceries, a cafe, a library) within a 15-minute walk, bike ride, or public transit ride of your home?

See https://github.com/dabreegster/abstreet/issues/393 for more context.

Structs

Entry

ExploreAmenities

HoverOnBuilding

Viewer
This is the UI state for exploring the isochrone/walkshed from a single building.

Functions

build_panel

draw_star

draw_unwalkable_roads

options_from_controls

options_to_controls

Type Definitions

HoverKey
(building, scale factor)