Struct fifteen_min::find_home::FindHome

pub struct FindHome {
    options: Options,
}

Ask what types of amenities are necessary to be within a walkshed, then rank every house with how many of those needs are satisfied.

Fields
options: Options
Implementations

impl FindHome

pub fn new(
ctx: &mut EventCtx<'_>,
options: Options
) -> Box<dyn State<SimpleApp<()>>>